+++ CHANGELOG.md
@@ -55,6 +55,8 @@
     * F1 key now is used again for the select tool
     * Fixed some menu actions that got removed or broken during the KDE transition
     * Fixed Alt-clicking on keyframes being inconsisent for the first keyframe
+    * Fixed keyframe multi-select not deselecting keyframes
+
 
 ## 0.6.0